+++ src/CartMule/CartMule/AppShell.xaml.cs
@@ -1,9 +1,12 @@
-﻿namespace CartMule;
+﻿using CartMule.Views;
+
+namespace CartMule;
 
 public partial class AppShell : Shell
 {
     public AppShell()
     {
         InitializeComponent();
+        Routing.RegisterRoute("listdetail", typeof(ListDetailPage));
     }
 }